MQTTHub/requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-dbrouter==0.1.1
django-filter==2.2.0
django-guardian==2.1.0
djangorestframework==3.10.3
djangorestframework-api-key==1.4.1
djangorestframework-simplejwt==4.3.0
idna==2.8
Markdown==3.1.1
paho-mqtt==1.5.0
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
